Sanity check the inode of our file handle hasn't changed during an expunge
We saw a nasty bug where IO::Async::Process was manually closing every
file descriptor including the Cache::FastMmap one, but then the cache
was still being held around. The mmaped region continued to work fine,
but something opened another file and re-used the existing file descriptor,
so locking appeared to work, but other processes were actually locking
a completely different underlying file.

Lets try and detect this insanity. We ideally would do it every time we
lock a page, but that's a lot for an incredibly rare edge case. Instead
just do it when we're about to do a page expunge. It should be enough to
catch the madness, but with pretty low overhead