Fix leaving locks when cache is initialised
Ouch, the refactoring that locks a page before initialising it to avoid
races was leaving pages locked, which was totally breaking caching with
multiple processes.

The mmc_lock_page and mmc_unlock_page APIs were messy. The first took
an offset, but the second read it from the cache structure. Instead
make both of them take an offset so it's more symmetric and obvious.